Add files via upload
Seattle. This property is in the heart of an expensive residential area on the west side of Lake Washington. Currently the property has a 3,000 square foot one story building that contains a convenience store (a small grocery store). Your friend will soon be the owner and operator of the store which is only moderately profitable. She is considering changing how this property is being utilized. Her strategy is to close the convenience store and renovate the building to open a coffee shop that also serves light healthy lunches and dinners. She has come to you for advice and is hoping you will agree to analyze this decision and provide a recommendation on whether she should close the convenience store and renovate the building to open the coffee shop. Assume the renovation will start at the end of June 2024 (time 0 is 6/30/2024). Additionally, your friend has stated that she is planning on living in this area for another 10 years (end of 2033) and is likely to sell the property at that time. Assume the sale price in 10 years will be the same regardless of the decision to renovate or not.

The revenue from the convenience store was $370,000 in 2022 and, based on performance to date, the forecasted revenue in 2023 is $380,000. After 2023, revenue is expected to grow by 2% per year in the future. The operating margin of the convenience store is expected to stay constant in the future at 8%, there is no depreciation expense, and operating working capital is historically 10% of revenue. 

The expected costs of renovating the building are $650,000 and it will take six months to complete the renovation (assume cash outflow for the renovation occurs at time 0). The renovation expense falls into the 10-year depreciable life. Assuming the renovation is completed by December 30th, 2024, the revenue from operating the coffee shop in 2025 is expected to equal $280,000. Revenue in 2026 is expected to equal $360,000 and revenue in 2027 is expected to equal $460,000. The expected annual growth in revenue is 6%/year from 2027 to 2031 and then revenue is expected to increase by 3% per year after 2031. The expected operating margin (excluding depreciation expense) is 30%, and the required operating working capital will be 5% of revenue. 

To finance the renovation your friend will invest $200,000 from her savings and has secured a bank loan for the remaining $450,000 at a rate of 8%/year. Assume, given the risk of this investment, that she will require a 12%/year return on her $200,000 equity investment. The expected income tax rate is 20% for both businesses. Assume your friend has enough income from her current job to benefit from any potential tax savings from losses created from the coffee shop in its early years. 

After estimating incremental after-tax cash flows from the renovation of the Madison Park property, calculate the NPV, IRR, and the Payback Period to provide a recommendation to your friend. Please provide a short memo written to her (no more than one or two pages) describing your process, results, and the implication of your analysis for your friend’s decision to renovate followed by your Excel spreadsheets (consider your audience, your friend, when constructing the memo). 

Note:  A MACRS Depreciation Schedule to be used to calculate depreciation expense from the renovation is provided on the following page.







Depreciation Expense in year t = Initial Outlay * Recovery Allowance Percentage in year t.      
Note: there is no adjustment for salvage value in the calculation of depreciation expense using the MACRS method of depreciation.
Also Note:  This MACRS schedule assumed a June 30 acquisition of the asset and, therefore, year 1’s allowance and the last year’s allowance are already adjusted to reflect depreciation expense for the half year.